+++ internal/maze/maze.go
@@ -131,9 +131,9 @@ func (p *Position) WithValue(value int) {
 	p.value = value
 }
 
-func (p *Position) Distance(q Position) int {
+func (p *Position) Distance(q Position) float64 {
 	x := p.row - q.row
 	y := p.col - q.col
 
-	return int(math.Sqrt(math.Pow(float64(x), 2) + math.Pow(float64(y), 2)))
+	return math.Sqrt(math.Pow(float64(x), 2) + math.Pow(float64(y), 2))
 }
